Update IRsendDemo.ino
As written this example will cause issues with some IR receivers. On the face of it it sends teh sony signal burst 3 times with a 40ms gap. However, it really continues to send the sony signal forever with a 40ms gap.

There needs to be a reasonable gap between signals sent & I have added in a 5 sec gap as a reasonable figure.

Without a gap, many IR receivers will treat this signal as noise as it send continuous sony bursts with a 40 ms gap.